Script0

+++ doubly_linked_lists/0-print_dlistint.c
@@ -1,5 +1,4 @@
 #include "lists.h"
-#include <stdio.h>
 
 /**
  * print_dlistint - imprimer tous les éléments d'une liste dlistint_t
@@ -9,7 +8,7 @@
 
 size_t print_dlistint(const dlistint_t *h)
 {
-	size_t count = 0;
+	int count = 0;
 
 	while (h != NULL)
 	{
